requirements.txt
Flask
python-dotenv
razorpay
Flask-SQLAlchemy
psycopg2-binary
Werkzeug
